+++ src/app/table/data-table/terra-data-table.component.ts
@@ -4,7 +4,8 @@ import {
     Output,
     ViewChild,
     EventEmitter,
-    OnInit
+    OnChanges,
+    SimpleChanges
 } from '@angular/core';
 import { Observable } from 'rxjs/Observable';
 import { TerraDataTableHeaderCellInterface } from './cell/terra-data-table-header-cell.interface';
@@ -26,12 +27,17 @@ import { TerraButtonInterface } from '../../button/data/terra-button.interface';
                styles:    [require('./terra-data-table.component.scss')],
                template:  require('./terra-data-table.component.html')
            })
-export class TerraDataTableComponent<S extends TerraBaseService, D extends TerraBaseData, I extends TerraPagerInterface>
+export class TerraDataTableComponent<S extends TerraBaseService, D extends TerraBaseData, I extends TerraPagerInterface> implements OnChanges
 {
     @ViewChild('viewChildHeaderCheckbox') viewChildHeaderCheckbox:TerraCheckboxComponent;
+    
     @Input() inputService:S;
     @Input() inputDataType:string;
+    @Input() inputHasCheckboxes:boolean;
+    @Input() inputHasInitialLoading:boolean;
+    
     @Output() outputDoPagingEvent = new EventEmitter<TerraPagerInterface>();
+    
     private _headerList:Array<TerraDataTableHeaderCellInterface>;
     private _rowList:Array<TerraDataTableRowInterface<D>>;
     private _selectedRowList:Array<TerraDataTableRowInterface<D>> = [];
@@ -42,12 +48,13 @@ export class TerraDataTableComponent<S extends TerraBaseService, D extends Terra
     private _defaultPagingSize:number;
     private _initialLoadingMessage:string;
     private _alert:TerraAlertComponent = TerraAlertComponent.getInstance();
+    private _langPrefix:string = 'terraDataTable';
+    
+    /**
+     * @deprecated
+     */
     @Input()
     private _hasCheckboxes:boolean;
-    @Input()
-    private _hasInitialLoading:boolean;
-    
-    private _langPrefix:string = 'terraDataTable';
     
     // Overlay
     //@ViewChild('viewChildOverlayDataTableSettings') viewChildOverlayDataTableSettings:TerraOverlayComponent;
@@ -58,8 +65,19 @@ export class TerraDataTableComponent<S extends TerraBaseService, D extends Terra
     
     constructor()
     {
-        this._hasInitialLoading = false;
         this._hasCheckboxes = true;
+        this.inputHasCheckboxes = true;
+        this.inputHasInitialLoading = false;
+    }
+    
+    
+    ngOnChanges(changes:SimpleChanges):void
+    {
+        if(changes['_hasCheckboxes'])
+        {
+            console.warn('_hasCheckboxes is deprecated. It will be removed in one of the upcoming releases. Please use inputHasCheckboxes instead.');
+            this.inputHasCheckboxes = changes['_hasCheckboxes'].currentValue;
+        }
     }
     
     private onHeaderCheckboxChange(isChecked:boolean):void